Implemented a working Newsfeed
