+++ .idea/misc.xml
@@ -9,25 +9,29 @@
         <entry key="..\:/Users/Hp/AndroidStudioProjects/WitsSocial/app/src/main/res/layout/activity_sign_in.xml" value="0.22" />
         <entry key="..\:/Users/Hp/AndroidStudioProjects/WitsSocial/app/src/main/res/layout/activity_sign_up.xml" value="0.28888888888888886" />
         <entry key="..\:/Users/Katleho Mphuthi/Documents/GitHub/Wits_Social/app/src/main/res/layout/activity_forgot_password.xml" value="0.140625" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1875" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable-v24/roundedcorner.xml" value="0.106" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/addphoto.xml" value="0.231" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/button_background.xml" value="0.231" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/clear.xml" value="0.2225" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/edit_text_background.xml" value="0.231" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/ic_baseline_add_photo_alternate_24.xml" value="0.231" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/ic_baseline_person_48.xml" value="0.2075" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/ic_launcher_background.xml" value="0.1875" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/ic_share_24.xml" value="0.1875" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/drawable/image_border.xml" value="0.106" />
-        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_check_mail.xml" value="0.18432971014492755" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_check_mail.xml" value="0.33" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_forgot_password.xml" value="0.18432971014492755" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_home.xml" value="0.18432971014492755" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_more.xml" value="0.2" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_more_options.xml" value="0.25" />
-        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_post.xml" value="0.18787878787878787" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_post.xml" value="0.24592391304347827" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_sign_in.xml" value="0.33" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_sign_up.xml" value="0.33" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_user_profile.xml" value="0.1841880341880342" />
-        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_welcome.xml" value="0.10916919959473151" />
-        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/fragment_home.xml" value="0.33" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/activity_welcome.xml" value="0.25" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/fragment_home.xml" value="0.22" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/fragment_newsfeed.xml" value="0.1" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/fragment_notification.xml" value="0.25" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/fragment_profile.xml" value="0.33" />
@@ -35,8 +39,8 @@
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/fragment_sign_out.xml" value="0.25" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/frame_textview.xml" value="0.18432971014492755" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/list_adapter.xml" value="0.14719202898550723" />
-        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/more_options_list_item.xml" value="0.12545289855072464" />
-        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/post_item.xml" value="0.18" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/more_options_list_item.xml" value="0.2" />
+        <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/post_item.xml" value="0.25" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/topbar.xml" value="0.22447916666666667" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/layout/user_item.xml" value="0.33" />
         <entry key="..\:/Users/Kgotso/Documents/COMS 3/COMS3009A_SD/Project/app/src/main/res/menu/bottom_menu.xml" value="0.22447916666666667" />
