+++ app/src/main/java/com/example/witssocial/Authentication/ForgotPasswordActivity.java
@@ -1,25 +1,22 @@
-package com.example.witssocial;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
+package com.example.witssocial.Authentication;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Patterns;
 import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.witssocial.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
 import com.google.firebase.auth.FirebaseAuth;
-
-import java.util.regex.Pattern;
 
 public class ForgotPasswordActivity extends AppCompatActivity {
 
@@ -66,7 +63,7 @@ public class ForgotPasswordActivity extends AppCompatActivity {
             public void onComplete(@NonNull Task<Void> task) {
                 if (task.isSuccessful())
                 {
-                    Intent intent = new Intent(ForgotPasswordActivity.this,check_mail.class);
+                    Intent intent = new Intent(ForgotPasswordActivity.this, check_mail.class);
                     startActivity(intent);
                 }else{
                     Toast.makeText(ForgotPasswordActivity.this, "Try again!", Toast.LENGTH_LONG).show();
