+++ app/src/main/java/com/example/witssocial/Authentication/SignUpActivity.java
@@ -1,4 +1,4 @@
-package com.example.witssocial;
+package com.example.witssocial.Authentication;
 
 
 import android.app.ProgressDialog;
@@ -11,6 +11,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.witssocial.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
