+++ app/src/main/java/com/example/witssocial/Home/HomeFragment.java
@@ -1,0 +1,101 @@
+package com.example.witssocial.Home;
+
+import android.content.Context;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.witssocial.Model.Post;
+import com.example.witssocial.Profile.UserProfileActivity;
+import com.example.witssocial.R;
+import com.example.witssocial.Utils.PostAdapter;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+import java.util.ArrayList;
+
+
+public class HomeFragment extends Fragment {
+    private static final String TAG = "HomeFragmentActivity";
+
+    DatabaseReference database;
+     ImageView profilePicture;
+  RecyclerView recyclerView;
+     PostAdapter postAdapter;
+    ArrayList<Post> list;
+
+    private FirebaseAuth mFirebaseAuth;
+    private static Context mContext;
+
+
+    public HomeFragment() {
+        // Required empty public constructor
+    }
+
+
+    public static HomeFragment newInstance(String param1, String param2) {
+        HomeFragment fragment = new HomeFragment();
+        Bundle args = new Bundle();
+
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+
+        View view =  inflater.inflate(R.layout.fragment_home, container, false);
+
+        profilePicture = view.findViewById(R.id.iv_home_profile_picture);
+
+        profilePicture.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getActivity(), UserProfileActivity.class);
+                startActivity(intent);
+            }
+        });
+
+        recyclerView = view.findViewById(R.id.recycler_view);
+        database = FirebaseDatabase.getInstance().getReference("Posts");
+        recyclerView.setHasFixedSize(true);
+        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+        list = new ArrayList<>();
+        postAdapter = new PostAdapter(getContext(), list);
+        recyclerView.setAdapter(postAdapter);
+
+       database.addValueEventListener(new ValueEventListener() {
+           @Override
+           public void onDataChange(@NonNull DataSnapshot snapshot) {
+               for (DataSnapshot dataSnapshot : snapshot.getChildren()){
+                   Post post = dataSnapshot.getValue(Post.class);
+                   list.add(post);
+               }
+
+               postAdapter.notifyDataSetChanged();
+           }
+
+           @Override
+           public void onCancelled(@NonNull DatabaseError error) {
+
+           }
+       });
+
+        return view;
+
+    }
+}
