+++ app/src/main/java/com/example/witssocial/Model/Post.java
@@ -2,12 +2,12 @@ package com.example.witssocial.Model;
 
 public class Post {
     private String image;
-    private String caption;
+    private String Caption;
     private String username;
 
     public Post(String image, String caption, String username) {
         this.image = image;
-        this.caption = caption;
+        this.Caption = caption;
         this.username = username;
     }
 
@@ -23,24 +23,17 @@ public class Post {
         return image;
     }
 
-    public void setImage(String image) {
-        this.image = image;
-    }
+
 
     public String getCaption() {
-        return caption;
+        return Caption;
     }
 
-    public void setCaption(String caption) {
-        this.caption = caption;
-    }
+
 
     public String getUsername() {
         return username;
     }
 
-    public void setUsername(String username) {
-        this.username = username;
-    }
 }
