+++ app/src/main/java/com/example/witssocial/Notification/NotificationFragment.java
@@ -1,12 +1,13 @@
-package com.example.witssocial;
+package com.example.witssocial.Notification;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.witssocial.R;
 
 /**
  * A simple {@link Fragment} subclass.
