--- app/src/main/java/com/example/witssocial/PostAdapter.java
@@ -1,114 +1,0 @@
-package com.example.witssocial;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.bumptech.glide.Glide;
-import com.example.witssocial.Model.Post;
-import com.example.witssocial.Model.User;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-
-import java.util.List;
-
-public class PostAdapter extends RecyclerView.Adapter<PostAdapter.ViewHolder>{
-
-    public Context mContext;
-    public List<Post> mPost;
-
-    private FirebaseUser firebaseUser;
-
-    public PostAdapter(Context context, List<Post> posts){
-        this.mContext = context;
-        this.mPost = posts;
-    }
-
-    @NonNull
-    @Override
-    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(mContext).inflate(R.layout.post_item, parent, false);
-        return new PostAdapter.ViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-
-        firebaseUser = FirebaseAuth.getInstance().getCurrentUser();
-        Post post = mPost.get(position);
-
-        Glide.with(mContext).load(post.getImage()).into(holder.post_image);
-
-        if (post.getCaption().equals("")){
-            holder.caption.setVisibility(View.GONE);
-        } else {
-            holder.caption.setVisibility(View.VISIBLE);
-            holder.caption.setText(post.getCaption());
-        }
-
-       publisherInfo(holder.username);
-
-
-    }
-
-    @Override
-    public int getItemCount() {
-        return mPost.size();
-    }
-
-    public class ViewHolder extends RecyclerView.ViewHolder {
-
-        public ImageView post_image;
-        public TextView username, caption;
-
-        public ViewHolder(@NonNull View itemView) {
-            super(itemView);
-
-            post_image = itemView.findViewById(R.id.post_image);
-            username = itemView.findViewById(R.id.username);
-            caption = itemView.findViewById(R.id.caption);
-        }
-    }
-
-    private void publisherInfo(final TextView username){
-        DatabaseReference reference = FirebaseDatabase.getInstance().getReference();
-
-        reference.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                /*
-                User user = snapshot.getValue(User.class);
-                Glide.with(mContext).load(user.getImageurl()).into(image_profile);
-                username.setText(user.getUsername());
-                */
-
-                for (DataSnapshot snapshot1 : snapshot.getChildren()){
-                    String rootNodeChild = snapshot1.getKey();
-                    //
-
-                    if(rootNodeChild.length() > 10){
-                        User user = snapshot1.getValue(User.class);
-                        username.setText(user.getUsername());
-                    }
-
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-}
