+++ app/src/main/java/com/example/witssocial/SignOut/SignOutFragment.java
@@ -12,8 +12,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
+import com.example.witssocial.Authentication.SignInActivity;
 import com.example.witssocial.R;
-import com.example.witssocial.SignInActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
