+++ app/src/main/java/com/example/witssocial/Utils/UserAdapter.java
@@ -1,4 +1,4 @@
-package com.example.witssocial;
+package com.example.witssocial.Utils;
 
 import android.content.Context;
 import android.content.Intent;
@@ -14,8 +14,10 @@ import androidx.fragment.app.FragmentActivity;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.example.witssocial.Home.MainActivity;
 import com.example.witssocial.Model.User;
 import com.example.witssocial.Profile.ProfileFragment;
+import com.example.witssocial.R;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
