+++ app/src/main/java/com/example/witssocial/Utils/Validator.java
@@ -1,0 +1,21 @@
+package com.example.witssocial.Utils;
+
+public class Validator {
+    String a;
+    public Validator(String b){
+        a = b;
+    }
+
+    public boolean isinputString(){
+        if(a instanceof String){
+            return true;
+        }else{
+            return false;
+        }
+    }
+
+    public int add(){
+        return 2;
+    }
+
+}
