+++ app/src/main/java/com/example/witssocial/WelcomeActivity.java
@@ -1,10 +1,13 @@
 package com.example.witssocial;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.witssocial.Authentication.SignInActivity;
+import com.example.witssocial.Authentication.SignUpActivity;
 
 public class WelcomeActivity extends AppCompatActivity {
 
@@ -15,10 +18,10 @@ public class WelcomeActivity extends AppCompatActivity {
     }
     public void goToSignUpActivity(View view) {
 
-        startActivity(new Intent(getApplicationContext(),SignUpActivity.class));
+        startActivity(new Intent(getApplicationContext(), SignUpActivity.class));
     }
 
     public void goToSignInActivity(View view) {
-        startActivity(new Intent(getApplicationContext(),SignInActivity.class));
+        startActivity(new Intent(getApplicationContext(), SignInActivity.class));
     }
 }
