+++ app/src/main/res/color/selector.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <item android:color="@color/black"
+    <item android:color="@color/md_theme_light_primary"
         android:state_selected="true"/>
 
-    <item android:color="@android:color/darker_gray"/>
+    <item android:color="@color/md_theme_light_onSurface"/>
 
 </selector>
