+++ app/src/main/res/drawable-v24/roundedcorner.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
     <shape xmlns:android="http://schemas.android.com/apk/res/android">
         <corners  android:radius="50dp"/>
-        <solid android:color="@color/purple_200"></solid>
+        <solid android:color="@color/material_on_surface_stroke"></solid>
     </shape>
