+++ app/src/main/res/drawable/edit_text_background.xml
@@ -5,6 +5,6 @@
     <corners android:radius="3dp"/>
     <solid android:color="#eee"/>
     <stroke android:width="1dp"
-        android:color="@color/colorAccent"/>
+        android:color="@color/md_theme_light_surface"/>
 
 </shape>
