+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForgotPasswordActivity"
+    tools:context=".Authentication.ForgotPasswordActivity"
     android:background="@color/white">
 
     <TextView
@@ -94,7 +94,6 @@
         android:layout_height="49dp"
         android:layout_marginTop="528dp"
         android:layout_marginBottom="72dp"
-        android:background="@drawable/roundedcorner"
         android:text="@string/submitbutton"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
