+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".Home.HomeActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
@@ -22,9 +22,9 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:id="@+id/bottom_navigation"
-            app:itemIconTint="@color/selector"
+            app:itemIconTint="?attr/colorOnSurfaceVariant"
+            app:backgroundTint="?attr/colorSurface"
             app:labelVisibilityMode="unlabeled"
-            android:background="?android:attr/windowBackground"
             app:menu="@menu/bottom_menu"/>
 
     </com.google.android.material.appbar.AppBarLayout>
