+++ app/src/main/res/layout/activity_main.xml
@@ -4,8 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/blue_700"
-    tools:context=".MainActivity">
+    android:background="@color/md_theme_light_primary"
+    tools:context=".Home.MainActivity">
 
     <ImageView
         android:id="@+id/imageView2"
