+++ app/src/main/res/layout/activity_welcome.xml
@@ -32,7 +32,7 @@
             android:text="Sign In"
             android:textAllCaps="false"
             app:cornerRadius="50dp"
-            app:rippleColor="@color/purple_500"
+            app:rippleColor="@color/md_theme_light_primary"
             app:strokeColor="#9900F0">
 
         </com.google.android.material.button.MaterialButton>
@@ -46,7 +46,7 @@
             android:text=" Sign Up"
             android:textAllCaps="false"
             app:cornerRadius="50dp"
-            app:strokeColor="@color/blue_700">
+            app:strokeColor="@color/md_theme_light_primary">
 
         </com.google.android.material.button.MaterialButton>
