+++ app/src/main/res/values/colors.xml
@@ -1,33 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="purple_200">#FFBB86FC</color>
-    <color name="purple_500">#FF6200EE</color>
-    <color name="purple_700">#FF3700B3</color>
-    <color name="teal_200">#FF03DAC5</color>
-    <color name="teal_700">#FF018786</color>
-    <color name="black">#FF000000</color>
-    <color name="white">#FFFFFFFF</color>
-    <color name="blue_700">#0057b8</color>
-    <color name="blue_500">#2d77c7</color>
-
-    <color name="colorPrimary">#455cde</color>
-    <color name="colorPrimaryDark">#000d3d</color>
-    <color name="colorAccent">#c2bdbf</color>
-    <color name="colorBlack">#000000</color>
-    <color name="colorWhite">#ffffff</color>
-
-    <color name="md_theme_light_primary">#1F45F7</color>
+    <color name="md_theme_light_primary">#284DE2</color>
     <color name="md_theme_light_onPrimary">#FFFFFF</color>
-    <color name="md_theme_light_primaryContainer">#DDE0FF</color>
-    <color name="md_theme_light_onPrimaryContainer">#000C62</color>
-    <color name="md_theme_light_secondary">#5B5D71</color>
+    <color name="md_theme_light_primaryContainer">#DDE1FF</color>
+    <color name="md_theme_light_onPrimaryContainer">#00105C</color>
+    <color name="md_theme_light_secondary">#5A5D72</color>
     <color name="md_theme_light_onSecondary">#FFFFFF</color>
-    <color name="md_theme_light_secondaryContainer">#E0E1FA</color>
+    <color name="md_theme_light_secondaryContainer">#DFE1FA</color>
     <color name="md_theme_light_onSecondaryContainer">#171A2C</color>
-    <color name="md_theme_light_tertiary">#77536D</color>
+    <color name="md_theme_light_tertiary">#76546E</color>
     <color name="md_theme_light_onTertiary">#FFFFFF</color>
-    <color name="md_theme_light_tertiaryContainer">#FFD7F3</color>
-    <color name="md_theme_light_onTertiaryContainer">#2D1228</color>
+    <color name="md_theme_light_tertiaryContainer">#FFD6F4</color>
+    <color name="md_theme_light_onTertiaryContainer">#2C1228</color>
     <color name="md_theme_light_error">#BA1B1B</color>
     <color name="md_theme_light_errorContainer">#FFDAD4</color>
     <color name="md_theme_light_onError">#FFFFFF</color>
@@ -36,26 +20,26 @@
     <color name="md_theme_light_onBackground">#1B1B1F</color>
     <color name="md_theme_light_surface">#FEFBFF</color>
     <color name="md_theme_light_onSurface">#1B1B1F</color>
-    <color name="md_theme_light_surfaceVariant">#E3E1EC</color>
-    <color name="md_theme_light_onSurfaceVariant">#46464F</color>
-    <color name="md_theme_light_outline">#767680</color>
+    <color name="md_theme_light_surfaceVariant">#E2E1EC</color>
+    <color name="md_theme_light_onSurfaceVariant">#45464F</color>
+    <color name="md_theme_light_outline">#76767F</color>
     <color name="md_theme_light_inverseOnSurface">#F3F0F5</color>
     <color name="md_theme_light_inverseSurface">#303034</color>
-    <color name="md_theme_light_inversePrimary">#B9C3FF</color>
+    <color name="md_theme_light_inversePrimary">#B8C3FF</color>
     <color name="md_theme_light_shadow">#000000</color>
-    <color name="md_theme_light_primaryInverse">#B9C3FF</color>
-    <color name="md_theme_dark_primary">#B9C3FF</color>
-    <color name="md_theme_dark_onPrimary">#001A98</color>
-    <color name="md_theme_dark_primaryContainer">#0029D3</color>
-    <color name="md_theme_dark_onPrimaryContainer">#DDE0FF</color>
-    <color name="md_theme_dark_secondary">#C4C5DD</color>
-    <color name="md_theme_dark_onSecondary">#2D2F42</color>
-    <color name="md_theme_dark_secondaryContainer">#434559</color>
-    <color name="md_theme_dark_onSecondaryContainer">#E0E1FA</color>
-    <color name="md_theme_dark_tertiary">#E5BAD7</color>
-    <color name="md_theme_dark_onTertiary">#44263D</color>
+    <color name="md_theme_light_primaryInverse">#B8C3FF</color>
+    <color name="md_theme_dark_primary">#B8C3FF</color>
+    <color name="md_theme_dark_onPrimary">#001F90</color>
+    <color name="md_theme_dark_primaryContainer">#0030C8</color>
+    <color name="md_theme_dark_onPrimaryContainer">#DDE1FF</color>
+    <color name="md_theme_dark_secondary">#C3C5DD</color>
+    <color name="md_theme_dark_onSecondary">#2C2F41</color>
+    <color name="md_theme_dark_secondaryContainer">#434659</color>
+    <color name="md_theme_dark_onSecondaryContainer">#DFE1FA</color>
+    <color name="md_theme_dark_tertiary">#E5BAD8</color>
+    <color name="md_theme_dark_onTertiary">#44263E</color>
     <color name="md_theme_dark_tertiaryContainer">#5D3C55</color>
-    <color name="md_theme_dark_onTertiaryContainer">#FFD7F3</color>
+    <color name="md_theme_dark_onTertiaryContainer">#FFD6F4</color>
     <color name="md_theme_dark_error">#FFB4A9</color>
     <color name="md_theme_dark_errorContainer">#930006</color>
     <color name="md_theme_dark_onError">#680003</color>
@@ -64,16 +48,12 @@
     <color name="md_theme_dark_onBackground">#E4E1E6</color>
     <color name="md_theme_dark_surface">#1B1B1F</color>
     <color name="md_theme_dark_onSurface">#E4E1E6</color>
-    <color name="md_theme_dark_surfaceVariant">#46464F</color>
+    <color name="md_theme_dark_surfaceVariant">#45464F</color>
     <color name="md_theme_dark_onSurfaceVariant">#C6C5CF</color>
     <color name="md_theme_dark_outline">#90909A</color>
     <color name="md_theme_dark_inverseOnSurface">#1B1B1F</color>
     <color name="md_theme_dark_inverseSurface">#E4E1E6</color>
-    <color name="md_theme_dark_inversePrimary">#1F45F7</color>
+    <color name="md_theme_dark_inversePrimary">#284DE2</color>
     <color name="md_theme_dark_shadow">#000000</color>
-    <color name="md_theme_dark_primaryInverse">#1F45F7</color>
-    <color name="seed">#1F45F7</color>
-    <color name="error">#BA1B1B</color>
-
-
+    <color name="md_theme_dark_primaryInverse">#284DE2</color>
 </resources>
