+++ app/src/main/res/values/themes.xml
@@ -1,36 +1,31 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Base application theme. -->
+<resources>
     <style name="Theme.WitsSocial" parent="Theme.Material3.Light.NoActionBar">
-
-        <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
-
-        <!-- Customize your theme here. -->
-
-
-        <!-- Original AppCompat attributes. -->
         <item name="colorPrimary">@color/md_theme_light_primary</item>
-        <item name="colorPrimaryDark">@color/md_theme_dark_primary</item>
+        <item name="colorOnPrimary">@color/md_theme_light_onPrimary</item>
+        <item name="colorPrimaryContainer">@color/md_theme_light_primaryContainer</item>
+        <item name="colorOnPrimaryContainer">@color/md_theme_light_onPrimaryContainer</item>
         <item name="colorSecondary">@color/md_theme_light_secondary</item>
-        <item name="textAppearanceHeadline1">@style/TextAppearance.MaterialComponents.Headline1</item>
-        <item name="textAppearanceHeadline2">@style/TextAppearance.MaterialComponents.Headline2</item>
-        <item name="textAppearanceHeadline3">@style/TextAppearance.MaterialComponents.Headline3</item>
-        <item name="textAppearanceHeadline4">@style/TextAppearance.MaterialComponents.Headline4</item>
-        <item name="textAppearanceHeadline5">@style/TextAppearance.MaterialComponents.Headline5</item>
-        <item name="textAppearanceHeadline6">@style/TextAppearance.MaterialComponents.Headline6</item>
-        <item name="textAppearanceSubtitle1">@style/TextAppearance.MaterialComponents.Subtitle1</item>
-        <item name="textAppearanceSubtitle2">@style/TextAppearance.MaterialComponents.Subtitle2</item>
-        <item name="textAppearanceBody1">@style/TextAppearance.MaterialComponents.Body1</item>
-        <item name="textAppearanceBody2">@style/TextAppearance.MaterialComponents.Body2</item>
-        <item name="textAppearanceCaption">@style/TextAppearance.MaterialComponents.Caption</item>
-        <item name="textAppearanceButton">@style/TextAppearance.MaterialComponents.Button</item>
-        <item name="textAppearanceOverline">@style/TextAppearance.MaterialComponents.Overline</item>
-
-        <!-- Material3 attributes (needed if parent="Theme.MaterialComponents"). -->
-
-
-
-
+        <item name="colorOnSecondary">@color/md_theme_light_onSecondary</item>
+        <item name="colorSecondaryContainer">@color/md_theme_light_secondaryContainer</item>
+        <item name="colorOnSecondaryContainer">@color/md_theme_light_onSecondaryContainer</item>
+        <item name="colorTertiary">@color/md_theme_light_tertiary</item>
+        <item name="colorOnTertiary">@color/md_theme_light_onTertiary</item>
+        <item name="colorTertiaryContainer">@color/md_theme_light_tertiaryContainer</item>
+        <item name="colorOnTertiaryContainer">@color/md_theme_light_onTertiaryContainer</item>
+        <item name="colorError">@color/md_theme_light_error</item>
+        <item name="colorErrorContainer">@color/md_theme_light_errorContainer</item>
+        <item name="colorOnError">@color/md_theme_light_onError</item>
+        <item name="colorOnErrorContainer">@color/md_theme_light_onErrorContainer</item>
+        <item name="android:colorBackground">@color/md_theme_light_background</item>
+        <item name="colorOnBackground">@color/md_theme_light_onBackground</item>
+        <item name="colorSurface">@color/md_theme_light_surface</item>
+        <item name="colorOnSurface">@color/md_theme_light_onSurface</item>
+        <item name="colorSurfaceVariant">@color/md_theme_light_surfaceVariant</item>
+        <item name="colorOnSurfaceVariant">@color/md_theme_light_onSurfaceVariant</item>
+        <item name="colorOutline">@color/md_theme_light_outline</item>
+        <item name="colorOnSurfaceInverse">@color/md_theme_light_inverseOnSurface</item>
+        <item name="colorSurfaceInverse">@color/md_theme_light_inverseSurface</item>
+        <item name="colorPrimaryInverse">@color/md_theme_light_primaryInverse</item>
         <item name="textAppearanceDisplayLarge">@style/TextAppearance.Material3.DisplayLarge</item>
         <item name="textAppearanceDisplayMedium">@style/TextAppearance.Material3.DisplayMedium</item>
         <item name="textAppearanceDisplaySmall">@style/TextAppearance.Material3.DisplaySmall</item>
@@ -50,40 +45,4 @@
         <item name="shapeAppearanceMediumComponent">@style/ShapeAppearance.Material3.MediumComponent</item>
         <item name="shapeAppearanceLargeComponent">@style/ShapeAppearance.Material3.LargeComponent</item>
     </style>
-
-        <!-- MaterialComponents attributes (needed if parent="Theme.AppCompat"). -->
-        <style name="AppTheme" parent="Theme.Material3.Light.NoActionBar">
-            <item name="colorPrimary">@color/md_theme_light_primary</item>
-            <item name="colorOnPrimary">@color/md_theme_light_onPrimary</item>
-            <item name="colorPrimaryContainer">@color/md_theme_light_primaryContainer</item>
-            <item name="colorOnPrimaryContainer">@color/md_theme_light_onPrimaryContainer</item>
-            <item name="colorSecondary">@color/md_theme_light_secondary</item>
-            <item name="colorOnSecondary">@color/md_theme_light_onSecondary</item>
-            <item name="colorSecondaryContainer">@color/md_theme_light_secondaryContainer</item>
-            <item name="colorOnSecondaryContainer">@color/md_theme_light_onSecondaryContainer</item>
-            <item name="colorTertiary">@color/md_theme_light_tertiary</item>
-            <item name="colorOnTertiary">@color/md_theme_light_onTertiary</item>
-            <item name="colorTertiaryContainer">@color/md_theme_light_tertiaryContainer</item>
-            <item name="colorOnTertiaryContainer">@color/md_theme_light_onTertiaryContainer</item>
-            <item name="colorError">@color/md_theme_light_error</item>
-            <item name="colorErrorContainer">@color/md_theme_light_errorContainer</item>
-            <item name="colorOnError">@color/md_theme_light_onError</item>
-            <item name="colorOnErrorContainer">@color/md_theme_light_onErrorContainer</item>
-            <item name="android:colorBackground">@color/md_theme_light_background</item>
-            <item name="colorOnBackground">@color/md_theme_light_onBackground</item>
-            <item name="colorSurface">@color/md_theme_light_surface</item>
-            <item name="colorOnSurface">@color/md_theme_light_onSurface</item>
-            <item name="colorSurfaceVariant">@color/md_theme_light_surfaceVariant</item>
-            <item name="colorOnSurfaceVariant">@color/md_theme_light_onSurfaceVariant</item>
-            <item name="colorOutline">@color/md_theme_light_outline</item>
-            <item name="colorOnSurfaceInverse">@color/md_theme_light_inverseOnSurface</item>
-            <item name="colorSurfaceInverse">@color/md_theme_light_inverseSurface</item>
-
-            <item name="colorPrimaryInverse">@color/md_theme_light_primaryInverse</item>
-        </style>
-
-
-
-
-
 </resources>
