+++ app/src/test/java/com/example/witssocial/ValidatorTest.java
@@ -1,11 +1,12 @@
 package com.example.witssocial;
 
+import static com.google.common.truth.Truth.assertThat;
+
+import com.example.witssocial.Utils.Validator;
+
 import org.junit.Test;
-import org.junit.internal.runners.JUnit38ClassRunner;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-
-import static com.google.common.truth.Truth.assertThat;
 
 
 @RunWith(JUnit4.class)
